pyproject.toml
[project]
name = "mrfr-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logging>=0.4.9.6",
    "nltk>=3.9.1",
    "openai>=1.98.0",
    "openai-cost-calculator>=1.0.2",
    "python-dotenv>=1.1.1",
]
